backend/requirements.txt
fastapi
uvicorn[standard]
huggingface_hub
pillow
numpy
tensorflow-cpu
